headless-server::input : create a default keyboard device
This change creates a default keyboard device when there is
no probed device from pepper-evdev. As the focused client will
be able to control the output(s) and will get key event(s) by
default, at least, we need to create a keyboard when there is no keyboard
device detected.

Change-Id: I7997e9c5cc2a94e85261f3d56d1975653e6097e9
Signed-off-by: Sung-Jin Park <sj76.park@samsung.com>